requirements.txt
   google-generativeai
   python-dotenv
   pytest
   opencv-python
   numpy